move feedback class to integration.ts, make isAnonymous work, hide false in createElement
